Consolidate _certified_snf.py and certified_snf/ module split (#1553)
Move the Smith normal form kernel from the sibling _certified_snf.py
into the certified_snf/ package as _kernel.py. Update all 3 importers
(test_inverse_unimodular.py, certified_snf/operations.py, and
topology/operations.py) to import from jacobian.domains.certified_snf._kernel
instead of jacobian.domains._certified_snf.
