+++ src/jacobian/domains/certified_snf/_kernel.py
