+++ src/jacobian/domains/certified_snf/operations.py
@@ -6,7 +6,7 @@ from jacobian.contracts.certified_snf import (
     CertifiedSmithNormalFormRequest,
     CertifiedSmithNormalFormResult,
 )
-from jacobian.domains._certified_snf import certificate_from_reduction, smith_reduce
+from jacobian.domains.certified_snf._kernel import certificate_from_reduction, smith_reduce
 from jacobian.domains._examples import example
 from jacobian.math_tools import MathTool, MathTools
